refactor(shared): import clamp from math, not the validation facade
The back-compat re-export made a 3-line numeric clamp drag the whole
validation/salvage/import pipeline (plus i18n) into 19 modules, and
closed the validation -> validationPlacement -> outlineFrame -> validation
import cycle. Every clamp import now points at its home module and the
re-export is gone.
