+++ src/core/cqrs/v2/domain/layout/setGridUnitMmY.ts
@@ -9,7 +9,7 @@
 
 import { z } from 'zod';
 import { ok, err, layoutInvalidOperation } from '@/core/result';
-import { clamp } from '@/shared/utils/validation';
+import { clamp } from '@/shared/utils/math';
 import { mm } from '@/core/types';
 import { GRID_PITCH_MM_MAX, gridPitchFloors } from '@/shared/utils/drawerOutline';
 import { defineCommand } from '../../defineCommand';
