+++ src/features/bin-inspector/hooks/useBinInspector.ts
@@ -18,7 +18,8 @@ import { useMutations } from '@/shared/contexts';
 import { calcMaxGridUnits, CONSTRAINTS, STAGING_ID } from '@/core/constants';
 import { getLayerZStartResult } from '@/shared/utils/collision';
 import { isOk, isErr, getUserMessage } from '@/core/result';
-import { clamp, canPlaceBin, validateCustomProperties } from '@/shared/utils/validation';
+import { canPlaceBin, validateCustomProperties } from '@/shared/utils/validation';
+import { clamp } from '@/shared/utils/math';
 import { isBinLocked, validateBinRotation } from '@/shared/utils/binLocation';
 import { expandPairIds } from '@/shared/utils/binPairs';
 import { mlTracking } from '@/shared/analytics/useMLTracking';
