+++ src/features/bin-inspector/hooks/useBinInspectorMultiActions.ts
@@ -19,7 +19,8 @@ import { CONSTRAINTS, STAGING_ID } from '@/core/constants';
 import { getLayerZStartResult } from '@/shared/utils/collision';
 import { isErr, isOk, type Result } from '@/core/result';
 import type { LayoutError } from '@/core/result';
-import { clamp, canPlaceBin } from '@/shared/utils/validation';
+import { canPlaceBin } from '@/shared/utils/validation';
+import { clamp } from '@/shared/utils/math';
 import { isBinLocked } from '@/shared/utils/binLocation';
 import { mlTracking } from '@/shared/analytics/useMLTracking';
 import type { TFunction } from '@/i18n/context';
